requirements.txt
streamlit
google-generativeai
# Add any other specific libraries if you used them